+++ src/ast/block/attribute/name.rs
@@ -11,7 +11,7 @@ use crate::ast::error::IllegalChar;
 pub struct AttributeName<'a>(Cow<'a, str>);
 
 impl<'a> AttributeName<'a> {
-    /// Try to create a new [`AttributeName`].
+    /// Attempts to convert a string to an [`AttributeName`].
     ///
     /// # Errors
     /// Will return an error if the string contains any of the following characters:
@@ -22,10 +22,8 @@ impl<'a> AttributeName<'a> {
     /// - `U+002F SOLIDUS (/)`
     /// - `U+003D EQUAL SIGN (=)`
     /// - `U+003E GREATER-THAN SIGN (>)`.
-    pub fn try_new(value: impl Into<Cow<'a, str>>) -> Result<Self, IllegalChar> {
-        let mut value = value.into();
-
-        if let Some(ch) = value.chars().find(|ch| {
+    pub fn from_str(src: &'a str) -> Result<Self, IllegalChar> {
+        if let Some(ch) = src.chars().find(|ch| {
             matches!(
                 ch,
                 '\u{0009}'
@@ -40,17 +38,17 @@ impl<'a> AttributeName<'a> {
             return Err(IllegalChar(ch));
         }
 
-        if value.contains(|ch: char| ch.is_ascii_uppercase()) {
-            value.to_mut().make_ascii_lowercase();
+        if src.contains(|ch: char| ch.is_ascii_uppercase()) {
+            Ok(Self(Cow::Owned(src.to_ascii_uppercase())))
+        } else {
+            Ok(Self(Cow::Borrowed(src)))
         }
-
-        Ok(Self(value))
     }
 
     /// Convert a string into an [`AttributeName`] **without** validating.
     pub unsafe fn from_str_unchecked(src: &'a str) -> Self {
         if cfg!(debug_assertions) {
-            match Self::try_new(src) {
+            match Self::from_str(src) {
                 Ok(val) => val,
                 Err(_) => {
                     panic!("AttributeName::from_str_unchecked() with illegal chars")
@@ -87,23 +85,16 @@ impl Display for AttributeName<'_> {
     }
 }
 
-impl<'a> TryFrom<Cow<'a, str>> for AttributeName<'a> {
-    type Error = IllegalChar;
-    fn try_from(value: Cow<'a, str>) -> Result<Self, Self::Error> {
-        Self::try_new(value)
-    }
-}
-
 impl<'a> TryFrom<&'a str> for AttributeName<'a> {
     type Error = IllegalChar;
     fn try_from(value: &'a str) -> Result<Self, Self::Error> {
-        Self::try_new(value)
+        Self::from_str(value)
     }
 }
 
-impl<'a> TryFrom<String> for AttributeName<'a> {
+impl<'a> TryFrom<&'a String> for AttributeName<'a> {
     type Error = IllegalChar;
-    fn try_from(value: String) -> Result<Self, Self::Error> {
-        Self::try_new(value)
+    fn try_from(value: &'a String) -> Result<Self, Self::Error> {
+        Self::from_str(&value)
     }
 }