+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `AttributeName`, `AttributeValue` and `BlockName` now implements `PartialOrd`, `Ord` and `Default`.
 - `Section` and `Block` now implements `PartialOrd`, `Ord` and `Hash`.
+- Renamed `AttributeName::try_new` to `from_str`.
 
 ### Fixed